requirements.txt
yfinance
requests
python-dotenv
notion-client
apscheduler
beautifulsoup4
google-genai
urllib3
groq
feedparser
